Fix compiling for MSVC

+++ src/image_tools/displayTools.hpp
@@ -3,7 +3,7 @@
 #include <array>
 #include <cstdint>
 #include <memory>
-#include <vector>
+#include <string>
 
 #include "../linalg/linalg.hpp"
 
@@ -61,9 +61,9 @@ public:
     std::tuple<uint8_t, uint8_t, uint8_t> toRGB()
     {
         const float gamma = 1.0f;
-        return std::make_tuple(UINT8(std::pow(x, gamma) * 255.0f),
-                               UINT8(std::pow(y, gamma) * 255.0f),
-                               UINT8(std::pow(z, gamma) * 255.0f));
+        return std::make_tuple(UINT8(powf(x, gamma) * 255.0f),
+                               UINT8(powf(y, gamma) * 255.0f),
+                               UINT8(powf(z, gamma) * 255.0f));
     }
 };
